WIP - player and server builds
