+++ Assets/Scripts/Client/ClientSyncState.cs
@@ -6,11 +6,14 @@ using System.Collections.Generic;
 using UnityEngine.SceneManagement;
 using ubv.common.data;
 
-namespace ubv
+namespace ubv.client.logic
 {
-    namespace client
+    abstract public class ClientSyncState
     {
-        namespace logic
+        protected udp.client.UDPClient m_client;
+        protected readonly object m_lock = new object();
+
+        public ClientSyncState(udp.client.UDPClient client)
         {
             abstract public class ClientSyncState
             {
@@ -32,7 +35,7 @@ namespace ubv
                 private int? m_playerID;
 
 #if NETWORK_SIMULATE
-                private bool m_playWithoutServer;
+        private bool m_playWithoutServer;
 #endif // NETWORK_SIMULATE
 
                 public ClientSyncInit( 
@@ -41,7 +44,7 @@ namespace ubv
                     string physicsScene, 
                     PlayerSettings playerSettings
 #if NETWORK_SIMULATE
-                    , ClientSync parent)
+            , ClientSync parent)
 #endif // NETWORK_SIMULATE
                 {
                     m_physicsScene = physicsScene;
@@ -57,15 +60,15 @@ namespace ubv
                     m_TCPClient.Subscribe(this);
 
 #if NETWORK_SIMULATE
-                    parent.ConnectButtonEvent.AddListener(SendConnectionRequestToServer);
-                    parent.PlayWithoutServerButtonEvent.AddListener(() => { m_playWithoutServer = true; });
+            parent.ConnectButtonEvent.AddListener(SendConnectionRequestToServer);
+            parent.PlayWithoutServerButtonEvent.AddListener(() => { m_playWithoutServer = true; });
 #endif // NETWORK_SIMULATE
-                }
+        }
 
-                public override ClientSyncState FixedUpdate()
-                {
-                    return this;
-                }
+        public override ClientSyncState FixedUpdate()
+        {
+            return this;
+        }
 
                 public void SendConnectionRequestToServer()
                 {
@@ -101,6 +104,8 @@ namespace ubv
                         }
                     }
                 }
+            }
+        }
 
                 public override ClientSyncState Update()
                 {
@@ -120,9 +125,9 @@ namespace ubv
                     }
 #endif // NETWORK_SIMULATE
 
-                    return this;
-                }
-            }
+            return this;
+        }
+    }
 
             /// <summary>
             /// Represents the state of the server during the game
@@ -132,23 +137,23 @@ namespace ubv
                 private udp.client.UDPClient m_UDPClient;
                 private readonly int m_playerID;
 
-                private uint m_remoteTick;
-                private uint m_localTick;
+        private uint m_remoteTick;
+        private uint m_localTick;
 
-                private const ushort CLIENT_STATE_BUFFER_SIZE = 64;
+        private const ushort CLIENT_STATE_BUFFER_SIZE = 64;
 
-                private ClientState[] m_clientStateBuffer;
-                private InputFrame[] m_inputBuffer;
-                private InputFrame m_lastInput;
-                private ClientState m_lastServerState;
+        private ClientState[] m_clientStateBuffer;
+        private InputFrame[] m_inputBuffer;
+        private InputFrame m_lastInput;
+        private ClientState m_lastServerState;
 
-                private PhysicsScene2D m_clientPhysics;
+        private PhysicsScene2D m_clientPhysics;
 
                 private List<IClientStateUpdater> m_updaters;
                 
 #if NETWORK_SIMULATE
-                private readonly float m_packetLossChance = 0.15f;
-                private readonly bool m_noServer;
+        private readonly float m_packetLossChance = 0.15f;
+        private readonly bool m_noServer;
 #endif // NETWORK_SIMULATE
 
                 public ClientSyncPlay(udp.client.UDPClient UDPClient, 
@@ -157,20 +162,20 @@ namespace ubv
                     PlayerSettings playerSettings,
                     List<PlayerState> playerStates
 #if NETWORK_SIMULATE
-                    , bool startWithoutServer = false
+            , bool startWithoutServer = false
 #endif // NETWORK_SIMULATE
                     )
                 {
 #if NETWORK_SIMULATE
-                    m_noServer = startWithoutServer;
+            m_noServer = startWithoutServer;
 #endif // NETWORK_SIMULATE
-                    m_playerID = playerID;
-                    m_localTick = 0;
-                    m_clientStateBuffer = new ClientState[CLIENT_STATE_BUFFER_SIZE];
-                    m_inputBuffer = new InputFrame[CLIENT_STATE_BUFFER_SIZE];
+            m_playerID = playerID;
+            m_localTick = 0;
+            m_clientStateBuffer = new ClientState[CLIENT_STATE_BUFFER_SIZE];
+            m_inputBuffer = new InputFrame[CLIENT_STATE_BUFFER_SIZE];
 
-                    m_clientPhysics = SceneManager.GetSceneByName(physicsScene).GetPhysicsScene2D();
-                    m_lastServerState = null;
+            m_clientPhysics = SceneManager.GetSceneByName(physicsScene).GetPhysicsScene2D();
+            m_lastServerState = null;
 
                     m_updaters = new List<IClientStateUpdater>();
 
@@ -190,9 +195,9 @@ namespace ubv
                         m_clientStateBuffer[i] = new ClientState();
                         m_clientStateBuffer[i].SetPlayerID(m_playerID);
                         
-                        foreach (PlayerState playerState in playerStates)
-                        {
-                            PlayerState player = new PlayerState(playerState);
+                foreach (PlayerState playerState in playerStates)
+                {
+                    PlayerState player = new PlayerState(playerState);
 
                             m_clientStateBuffer[i].AddPlayer(player);
                         }
@@ -200,28 +205,32 @@ namespace ubv
                         m_inputBuffer[i] = new InputFrame();
                     }
                 }
+                        
+                m_inputBuffer[i] = new common.data.InputFrame();
+            }
+        }
 
-                public override ClientSyncState FixedUpdate()
-                {
-                    uint bufferIndex = m_localTick % CLIENT_STATE_BUFFER_SIZE;
+        public override ClientSyncState FixedUpdate()
+        {
+            uint bufferIndex = m_localTick % CLIENT_STATE_BUFFER_SIZE;
 
-                    UpdateInput(bufferIndex);
+            UpdateInput(bufferIndex);
 
-                    UpdateClientState(bufferIndex);
+            UpdateClientState(bufferIndex);
 
-                    ++m_localTick;
+            ++m_localTick;
 
-                    ClientCorrection();
+            ClientCorrection();
 
-                    return this;
-                }
+            return this;
+        }
 
                 public override ClientSyncState Update()
                 {
                     m_lastInput = InputController.CurrentFrame();
 
-                    return this;
-                }
+            return this;
+        }
 
                 public void RegisterUpdater(IClientStateUpdater updater)
                 {
@@ -255,57 +264,50 @@ namespace ubv
 
                 public void ReceivePacket(UDPToolkit.Packet packet)
                 {
-                    // TODO remove tick from ClientSTate and add it to custom server state packet?
-                    // client doesnt need its own client state ticks
-                    lock (m_lock)
-                    {
-                        ClientState state = udp.Serializable.FromBytes<ClientState>(packet.Data);
-                        if (state != null)
-                        {
-                            state.SetPlayerID(m_playerID);
-                            m_lastServerState = state;
+                    state.SetPlayerID(m_playerID);
+                    m_lastServerState = state;
 #if DEBUG_LOG
-                            Debug.Log("Received server state tick " + state.Tick);
+                    Debug.Log("Received server state tick " + state.Tick);
 #endif //DEBUG_LOG
-                            m_remoteTick = state.Tick;
-                        }
-                    }
+                    m_remoteTick = state.Tick;
                 }
+            }
+        }
 
-                private void UpdateInput(uint bufferIndex)
-                {
-                    if (m_lastInput != null)
-                    {
-                        m_inputBuffer[bufferIndex].Movement.Set(m_lastInput.Movement.Value);
-                        m_inputBuffer[bufferIndex].Sprinting.Set(m_lastInput.Sprinting.Value);
-                    }
-                    else
-                    {
-                        m_inputBuffer[bufferIndex].SetToNeutral();
-                    }
+        private void UpdateInput(uint bufferIndex)
+        {
+            if (m_lastInput != null)
+            {
+                m_inputBuffer[bufferIndex].Movement.Set(m_lastInput.Movement.Value);
+                m_inputBuffer[bufferIndex].Sprinting.Set(m_lastInput.Sprinting.Value);
+            }
+            else
+            {
+                m_inputBuffer[bufferIndex].SetToNeutral();
+            }
 
-                    m_inputBuffer[bufferIndex].Tick.Set(m_localTick);
+            m_inputBuffer[bufferIndex].Tick.Set(m_localTick);
 
-                    m_lastInput = null;
+            m_lastInput = null;
 
 
 #if NETWORK_SIMULATE
-                    if (m_noServer)
-                        return;
+            if (m_noServer)
+                return;
 #endif // NETWORK_SIMULATE
-                    // TODO: Cap max input queue size
-                    // (under the hood, send multiple packets?)
-                    List<common.data.InputFrame> frames = new List<common.data.InputFrame>();
-                    for (uint tick = m_remoteTick; tick <= m_localTick; tick++)
-                    {
-                        frames.Add(m_inputBuffer[tick % CLIENT_STATE_BUFFER_SIZE]);
-                    }
+            // TODO: Cap max input queue size
+            // (under the hood, send multiple packets?)
+            List<common.data.InputFrame> frames = new List<common.data.InputFrame>();
+            for (uint tick = m_remoteTick; tick <= m_localTick; tick++)
+            {
+                frames.Add(m_inputBuffer[tick % CLIENT_STATE_BUFFER_SIZE]);
+            }
 
                     InputMessage inputMessage = new InputMessage();
 
-                    inputMessage.PlayerID.Set(m_playerID);
-                    inputMessage.StartTick.Set(m_remoteTick);
-                    inputMessage.InputFrames.Set(frames);
+            inputMessage.PlayerID.Set(m_playerID);
+            inputMessage.StartTick.Set(m_remoteTick);
+            inputMessage.InputFrames.Set(frames);
 
 #if NETWORK_SIMULATE
                     if (Random.Range(0f, 1f) > m_packetLossChance)
@@ -317,10 +319,10 @@ namespace ubv
                         Debug.Log("SIMULATING PACKET LOSS");
                     }
 #else
-                    m_udpClient.Send(inputMessage.ToBytes());
+            m_udpClient.Send(inputMessage.ToBytes());
 #endif //NETWORK_SIMULATE       
                     
-                }
+        }
 
                 private void UpdateClientState(uint bufferIndex)
                 {
@@ -331,17 +333,22 @@ namespace ubv
                         Time.fixedDeltaTime);
                 }
 
-                private void ClientCorrection()
-                {
+        private void ClientCorrection()
+        {
 #if NETWORK_SIMULATE
-                    if (m_noServer)
-                        return;
+            if (m_noServer)
+                return;
 #endif // NETWORK_SIMULATE
-                    // receive a state from server
-                    // check what tick it corresponds to
-                    // rewind client state to the tick
-                    // replay up to local tick by stepping every tick
-                    lock (m_lock)
+            // receive a state from server
+            // check what tick it corresponds to
+            // rewind client state to the tick
+            // replay up to local tick by stepping every tick
+            lock (m_lock)
+            {
+                if (m_lastServerState != null)
+                {
+                    List<IClientStateUpdater> updaters = ClientState.UpdatersNeedingCorrection(m_lastServerState);
+                    for (int i = 0; i < updaters.Count; i++)
                     {
                         if (m_lastServerState != null)
                         {
@@ -350,24 +357,22 @@ namespace ubv
                             {
                                 uint rewindTicks = m_remoteTick;
 
-                                // reset world state to last server-sent state
-                                updaters[i].UpdateFromState(m_lastServerState);
+                        // reset world state to last server-sent state
+                        updaters[i].UpdateFromState(m_lastServerState);
 
-                                while (rewindTicks < m_localTick)
-                                {
-                                    uint rewindIndex = rewindTicks++ % CLIENT_STATE_BUFFER_SIZE;
+                        while (rewindTicks < m_localTick)
+                        {
+                            uint rewindIndex = rewindTicks++ % CLIENT_STATE_BUFFER_SIZE;
 
-                                    updaters[i].SetStateAndStep(
-                                        ref m_clientStateBuffer[rewindIndex],
-                                        m_inputBuffer[rewindIndex],
-                                        Time.fixedDeltaTime);
-                                    m_clientPhysics.Simulate(Time.fixedDeltaTime);
-                                }
-                            }
-
-                            m_lastServerState = null;
+                            updaters[i].SetStateAndStep(
+                                ref m_clientStateBuffer[rewindIndex],
+                                m_inputBuffer[rewindIndex],
+                                Time.fixedDeltaTime);
+                            m_clientPhysics.Simulate(Time.fixedDeltaTime);
                         }
                     }
+
+                    m_lastServerState = null;
                 }
             }
         }
