+++ Assets/Scripts/Common/InputMessage.cs
@@ -1,57 +1,51 @@
 ﻿using UnityEngine;
 using System.Collections.Generic;
 
-namespace ubv
+namespace ubv.common.data
 {
-    namespace common
+    public class InputFrame : serialization.Serializable
     {
-        namespace data
+        public serialization.types.Bool Sprinting;
+        public serialization.types.Vector2 Movement;
+        public serialization.types.Uint32 Tick;
+
+        public void SetToNeutral()
         {
-            public class InputFrame : udp.Serializable
-            {
-                public udp.SerializableTypes.Bool Sprinting;
-                public udp.SerializableTypes.Vector2 Movement;
-                public udp.SerializableTypes.Uint32 Tick;
+            Movement.Set(Vector2.zero);
+            Sprinting.Set(false);
+        }
 
-                public void SetToNeutral()
-                {
-                    Movement.Set(Vector2.zero);
-                    Sprinting.Set(false);
-                }
+        protected override void InitSerializableMembers()
+        {
+            Sprinting = new serialization.types.Bool(this, false);
+            Movement = new serialization.types.Vector2(this, Vector2.zero);
+            Tick = new serialization.types.Uint32(this, 0);
 
-                protected override void InitSerializableMembers()
-                {
-                    Sprinting = new udp.SerializableTypes.Bool(this, false);
-                    Movement = new udp.SerializableTypes.Vector2(this, Vector2.zero);
-                    Tick = new udp.SerializableTypes.Uint32(this, 0);
+            SetToNeutral();
+        }
 
-                    SetToNeutral();
-                }
+        protected override byte SerializationID()
+        {
+            return (byte)serialization.ID.INPUT_FRAME;
+        }
+    }
 
-                protected override byte SerializationID()
-                {
-                    return (byte)udp.Serialization.BYTE_TYPE.INPUT_FRAME;
-                }
-            }
+    public class InputMessage : serialization.Serializable
+    {
+        public serialization.types.Int32 PlayerID;
+        public serialization.types.Uint32 StartTick;
+        public serialization.types.List<InputFrame> InputFrames;
 
-            public class InputMessage : udp.Serializable
-            {
-                public udp.SerializableTypes.Int32 PlayerID;
-                public udp.SerializableTypes.Uint32 StartTick;
-                public udp.SerializableTypes.List<InputFrame> InputFrames;
+        protected override void InitSerializableMembers()
+        {
+            StartTick = new serialization.types.Uint32(this, 0);
+            InputFrames = new serialization.types.List<InputFrame>(this, new List<InputFrame>());
+            PlayerID = new serialization.types.Int32(this, 0);
+        }
 
-                protected override void InitSerializableMembers()
-                {
-                    StartTick = new udp.SerializableTypes.Uint32(this, 0);
-                    InputFrames = new udp.SerializableTypes.List<InputFrame>(this, new List<InputFrame>());
-                    PlayerID = new udp.SerializableTypes.Int32(this, 0);
-                }
-
-                protected override byte SerializationID()
-                {
-                    return (byte)udp.Serialization.BYTE_TYPE.INPUT_MESSAGE;
-                }
-            }
+        protected override byte SerializationID()
+        {
+            return (byte)serialization.ID.INPUT_MESSAGE;
         }
     }
 }
