+++ Assets/Scripts/Common/ScriptableObjects/StandardMovementSettings.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using UnityEditor;
+
+namespace ubv.common
+{
+    [CreateAssetMenu(fileName = "StandardMovementSettings", menuName = "ScriptableObjects/Settings/StandardMovementSettings", order = 1)]
+    public class StandardMovementSettings : ScriptableObject
+    {
+        public float WalkVelocity;
+        public float SprintVelocity;
+        public float Acceleration;
+    }
+}
