+++ Assets/Scripts/Serialization.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0c3c8895889320349a6089818136fdff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
