+++ Assets/Scripts/Serialization/Serializable.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 62af5f924ecd511429a3a6521f7e84c9
+guid: e77674170f5fc51479ce7f048feff927
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
