+++ Assets/Scripts/Serialization/Serialization.cs
@@ -1,0 +1,12 @@
+﻿namespace ubv.serialization
+{
+    public enum ID : byte
+    {
+        INPUT_FRAME = 0x00,
+        INPUT_MESSAGE = 0x01,
+        CLIENT_STATE = 0x02,
+        PLAYER_STATE = 0x03,
+        AUTH_MESSAGE = 0x04,
+        START_MESSAGE = 0x05
+    }
+}
