+++ Assets/Scripts/Server/ServerState.cs
@@ -5,11 +5,14 @@ using ubv.udp;
 using System.Collections.Generic;
 using ubv.tcp;
 
-namespace ubv
+namespace ubv.server.logic
 {
-    namespace server
+    abstract public class ServerState
     {
-        namespace logic
+        protected udp.server.UDPServer m_server;
+        protected readonly object m_lock = new object();
+
+        public ServerState(udp.server.UDPServer server)
         {
             abstract public class ServerState
             {
@@ -19,21 +22,29 @@ namespace ubv
                 public abstract ServerState FixedUpdate();
             }
 
+    public class ClientConnection
+    {
+        public uint ServerTick;
+        public client.ClientState State;
 
-            public class ClientConnection
-            {
-                public uint ServerTick;
-                public client.ClientState State;
+        public int PlayerGUID { get; private set; }
 
-                public int PlayerGUID { get; private set; }
+        public ClientConnection(int playerGUID)
+        {
+            State = new client.ClientState();
+            State.SetPlayerID(playerGUID);
+            PlayerGUID = playerGUID;
+        }
+    }
 
-                public ClientConnection(int playerGUID)
-                {
-                    State = new client.ClientState();
-                    State.SetPlayerID(playerGUID);
-                    PlayerGUID = playerGUID;
-                }
-            }
+    /// <summary>
+    /// Represents the state of the server before the game.
+    /// In charge of regrouping player parties, and launching 
+    /// the game with a fixed number of players
+    /// </summary>
+    public class GameCreationState : ServerState, udp.server.IServerReceiver
+    {
+        private Dictionary<IPEndPoint, ClientConnection> m_clientConnections;
 
             /// <summary>
             /// Represents the state of the server before the game.
@@ -53,10 +64,8 @@ namespace ubv
                 private readonly int m_snapshotDelay;
                 private readonly GameObject m_playerPrefab;
 
-                private List<common.data.PlayerState> m_players;
-
 #if NETWORK_SIMULATE
-                private bool m_forceStartGame;
+        private bool m_forceStartGame;
 #endif // NETWORK_SIMULATE
 
                 public GameCreationState(udp.server.UDPServer UDPServer, 
@@ -67,7 +76,7 @@ namespace ubv
                     int snapshotDelay, 
                     string physicsScene
 #if NETWORK_SIMULATE
-                    , ServerUpdate parent
+            , ServerUpdate parent
 #endif // NETWORK_SIMULATE 
                     )
                 {
@@ -77,15 +86,15 @@ namespace ubv
                     m_TCPClientConnections = new Dictionary<IPEndPoint, ClientConnection>();
                     m_UDPClientConnections = new Dictionary<IPEndPoint, ClientConnection>();
 
-                    m_movementSettings = movementSettings;
-                    m_snapshotDelay = snapshotDelay;
-                    m_physicsScene = physicsScene;
-                    m_playerPrefab = playerPrefab;
+            m_movementSettings = movementSettings;
+            m_snapshotDelay = snapshotDelay;
+            m_physicsScene = physicsScene;
+            m_playerPrefab = playerPrefab;
 
-                    m_forceStartGame = false;
+            m_forceStartGame = false;
 
 #if NETWORK_SIMULATE
-                    parent.ForceStartGameButtonEvent.AddListener(() => { m_forceStartGame = true; });
+            parent.ForceStartGameButtonEvent.AddListener(() => { m_forceStartGame = true; });
 #endif // NETWORK_SIMULATE 
                     
                     m_TCPServer.Subscribe(this);
@@ -97,7 +106,7 @@ namespace ubv
                     {
                         if (m_TCPClientConnections.Count > 3
 #if NETWORK_SIMULATE
-                            || m_forceStartGame
+                    || m_forceStartGame
 #endif // NETWORK_SIMULATE
                         )
                         {
@@ -115,13 +124,17 @@ namespace ubv
                             return new GameplayState(m_UDPserver, m_playerPrefab, m_UDPClientConnections, m_movementSettings, m_snapshotDelay, m_physicsScene);
                         }
                     }
-                    return this;
-                }
 
-                public override ServerState FixedUpdate()
-                {
-                    return this;
+                    return new GameplayState(m_server, m_playerPrefab, m_clientConnections, m_movementSettings, m_snapshotDelay, m_physicsScene);
                 }
+            }
+            return this;
+        }
+
+        public override ServerState FixedUpdate()
+        {
+            return this;
+        }
                 
                 public void ReceivePacket(TCPToolkit.Packet packet, IPEndPoint clientIP)
                 {
@@ -140,10 +153,10 @@ namespace ubv
                         common.data.IdentificationMessage idMessage = new common.data.IdentificationMessage();
                         idMessage.PlayerID.Set(playerID);
 
-                        common.data.PlayerState playerState = new common.data.PlayerState();
-                        playerState.GUID.Set(playerID);
+                common.data.PlayerState playerState = new common.data.PlayerState();
+                playerState.GUID.Set(playerID);
 
-                        // set rotation / position according to existing players?
+                // set rotation / position according to existing players?
 
                         m_players.Add(playerState);
                         m_UDPserver.Subscribe(this);
@@ -182,6 +195,7 @@ namespace ubv
                     }
                 }
             }
+        }
 
             /// <summary>
             /// Represents the state of the server during the game
@@ -194,14 +208,14 @@ namespace ubv
                 private Dictionary<ClientConnection, common.data.InputMessage> m_clientInputs;
                 private Dictionary<int, Rigidbody2D> m_bodies;
                 
-                private common.StandardMovementSettings m_movementSettings;
-                private readonly int m_snapshotDelay;
+        private common.StandardMovementSettings m_movementSettings;
+        private readonly int m_snapshotDelay;
                 
-                private uint m_tickAccumulator;
+        private uint m_tickAccumulator;
 
-                private PhysicsScene2D m_serverPhysics;
+        private PhysicsScene2D m_serverPhysics;
 
-                private GameObject m_playerPrefab;
+        private GameObject m_playerPrefab;
 
                 public GameplayState(udp.server.UDPServer UDPServer,
                     GameObject playerPrefab, 
@@ -215,10 +229,10 @@ namespace ubv
                     m_tickAccumulator = 0;
                     m_UDPClientConnections = UDPClientConnections;
 
-                    m_movementSettings = movementSettings;
+            m_movementSettings = movementSettings;
 
-                    m_serverPhysics = UnityEngine.SceneManagement.SceneManager.GetSceneByName(physicsScene).GetPhysicsScene2D();
-                    m_playerPrefab = playerPrefab;
+            m_serverPhysics = UnityEngine.SceneManagement.SceneManager.GetSceneByName(physicsScene).GetPhysicsScene2D();
+            m_playerPrefab = playerPrefab;
 
                     m_bodies = new Dictionary<int, Rigidbody2D>();
                     m_clientInputs = new Dictionary<ClientConnection, common.data.InputMessage>();
@@ -260,58 +274,62 @@ namespace ubv
 
                 public override ServerState Update()
                 {
-                    return this;
+                    common.data.PlayerState player = new common.data.PlayerState();
+                    player.GUID.Set(id);
+                    m_clientConnections[ip].State.AddPlayer(player);
+                    m_clientConnections[ip].State.SetPlayerID(id);
                 }
+            }
+        }
 
-                public override ServerState FixedUpdate()
+        public override ServerState Update()
+        {
+            return this;
+        }
+
+        public override ServerState FixedUpdate()
+        {
+            // for each player
+            // check if missing frames
+            // update frames
+
+            uint framesToSimulate = 0;
+            lock (m_lock)
+            {
+                foreach (ClientConnection client in m_clientInputs.Keys)
                 {
-                    // for each player
-                    // check if missing frames
-                    // update frames
-
-                    uint framesToSimulate = 0;
-                    lock (m_lock)
-                    {
-                        foreach (ClientConnection client in m_clientInputs.Keys)
-                        {
-                            common.data.InputMessage message = m_clientInputs[client];
-                            int messageCount = message.InputFrames.Value.Count;
-                            uint maxTick = message.StartTick + (uint)(messageCount - 1);
+                    common.data.InputMessage message = m_clientInputs[client];
+                    int messageCount = message.InputFrames.Value.Count;
+                    uint maxTick = message.StartTick + (uint)(messageCount - 1);
 #if DEBUG_LOG
-            Debug.Log("max tick to simulate = " + maxTick.ToString());
+    Debug.Log("max tick to simulate = " + maxTick.ToString());
 #endif // DEBUG_LOG
 
-                            // on recule jusqu'à ce qu'on trouve le  tick serveur le plus récent
-                            uint missingFrames = (maxTick > client.ServerTick) ? maxTick - client.ServerTick : 0;
+                    // on recule jusqu'à ce qu'on trouve le  tick serveur le plus récent
+                    uint missingFrames = (maxTick > client.ServerTick) ? maxTick - client.ServerTick : 0;
 
-                            if (framesToSimulate < missingFrames) framesToSimulate = missingFrames;
-                        }
+                    if (framesToSimulate < missingFrames) framesToSimulate = missingFrames;
+                }
 
-                        for (uint f = framesToSimulate; f > 0; f--)
+                for (uint f = framesToSimulate; f > 0; f--)
+                {
+                    foreach (ClientConnection client in m_clientInputs.Keys)
+                    {
+                        common.data.InputMessage message = m_clientInputs[client];
+                        int messageCount = message.InputFrames.Value.Count;
+                        if (messageCount > f)
                         {
-                            foreach (ClientConnection client in m_clientInputs.Keys)
-                            {
-                                common.data.InputMessage message = m_clientInputs[client];
-                                int messageCount = message.InputFrames.Value.Count;
-                                if (messageCount > f)
-                                {
-                                    common.data.InputFrame frame = message.InputFrames.Value[messageCount - (int)f - 1];
+                            common.data.InputFrame frame = message.InputFrames.Value[messageCount - (int)f - 1];
 
-                                    // must be called in main unity thread
-                                    Rigidbody2D body = m_bodies[client.PlayerGUID];
+                            // must be called in main unity thread
+                            Rigidbody2D body = m_bodies[client.PlayerGUID];
 
-                                    common.logic.PlayerMovement.Execute(ref body, m_movementSettings, frame, Time.fixedDeltaTime);
+                            common.logic.PlayerMovement.Execute(ref body, m_movementSettings, frame, Time.fixedDeltaTime);
 
-                                    client.State.GetPlayer().Position.Set(body.position);
-                                    client.State.Tick.Set(client.ServerTick);
-                                    client.ServerTick++;
-                                }
-                            }
-
-                            m_serverPhysics.Simulate(Time.fixedDeltaTime);
+                            client.State.GetPlayer().Position.Set(body.position);
+                            client.State.Tick.Set(client.ServerTick);
+                            client.ServerTick++;
                         }
-
-                        m_clientInputs.Clear();
                     }
 
                     if (++m_tickAccumulator > m_snapshotDelay)
@@ -325,7 +343,13 @@ namespace ubv
                     return this;
                 }
 
-                public void Receive(udp.UDPToolkit.Packet packet, IPEndPoint clientEndPoint)
+                m_clientInputs.Clear();
+            }
+
+            if (++m_tickAccumulator > m_snapshotDelay)
+            {
+                m_tickAccumulator = 0;
+                foreach (IPEndPoint ip in m_clientConnections.Keys)
                 {
                     common.data.InputMessage inputs = udp.Serializable.FromBytes<common.data.InputMessage>(packet.Data);
                     if (inputs != null && m_UDPClientConnections.ContainsKey(clientEndPoint))
@@ -337,6 +361,16 @@ namespace ubv
                     }
                 }
             }
+        }
+
+        public void OnConnect(IPEndPoint clientIP)
+        {
+            //throw new System.NotImplementedException();
+        }
+                
+        public void OnDisconnect(IPEndPoint clientIP)
+        {
+            //m_IPConnections.Remove(clientIP);
         }
     }
 }
